refactor: extract title disambiguation into _DisambiguationMixin
Sixth decomposition cut (~225 lines). Moves the contiguous disambiguation
block — _is_disambig_twin_path, _probe_disambig_twin,
_auto_pick_canonical_over_disambig_twin, _auto_pick_canonical_over_extends_topic,
_render_disambiguation — into openzim_mcp/disambiguation.py as
_DisambiguationMixin.

External dep is just zim_operations (declared under TYPE_CHECKING).
_render_disambiguation was a staticmethod that hard-referenced
`SimpleToolsHandler._recase_from_original` by class name; converted to a
classmethod using `cls._recase_from_original` (declared under TYPE_CHECKING)
— transparent to its single `self._render_disambiguation(...)` caller. Also
dropped a stale docstring-only `is_strong_title_match` import.

Pure code movement, no behavior change. simple_tools.py 3,966 → 3,744
(5,896 → 3,744 across six mixin cuts; -2,152 lines / -36%). Full suite green
(2623 passed); mypy/flake8/isort clean.
